// src/protocol/types.ts
export interface MatchMakerSetup {
  PossibleQueues?: MatchMakerSetup_Queue[];
}

export interface MatchMakerSetup_Queue {
  Name?: string;
  Description?: string;
  Maps?: string[];
  Game?: string;
  MaxPartySize: number;
}

export interface MatchMakerQueueRequest {
  Queues?: string[];
}

export interface MatchMakerStatus {
  JoinedQueues?: string[];
  QueueCounts?: Record<string, number>;
  CurrentEloWidth?: number;
  JoinedTime?: string;
  BannedSeconds?: number;
  InstantStartQueues?: string[];
  IngameCounts?: Record<string, number>;
  UserCount: number;
  UserCountDiscord: number;
}

export interface AreYouReady {
  MinimumWinChance: number;
  QuickPlay: boolean;
  SecondsRemaining: number;
}

export interface AreYouReadyUpdate {
  ReadyAccepted: boolean;
  LikelyToPlay: boolean;
  QueueReadyCounts?: Record<string, number>;
  YourBattleSize?: number;
  YourBattleReady?: number;
}

export interface AreYouReadyResult {
  IsBattleStarting: boolean;
  AreYouBanned: boolean;
}